+++ qbraid/programs/_import.py
@@ -108,6 +108,8 @@ def _get_class(module: str):
         return qibo.Circuit  # type: ignore # noqa: F821
     if module == "stim":  # pragma: no cover
         return stim.Circuit  # type: ignore # noqa: F821
+    if module == "pulser":
+        return pulser.sequence.sequence.Sequence  # type: ignore # noqa: F821
     raise ValueError(f"Unsupported module '{module}'")
 
 
@@ -129,7 +131,7 @@ dynamic_type_registry: dict[str, Type[Any]] = _dynamic_importer(
     ]
 )
 dynamic_non_native: dict[str, Type[Any]] = _dynamic_importer(
-    ["bloqade.analog.builder.assign", "qibo", "stim", "pyqir"]
+    ["bloqade.analog.builder.assign", "qibo", "stim", "pyqir", "pulser"]
 )
 static_type_registry: dict[str, Type[Any]] = {
     metatype.__alias__: metatype.__bound__ for metatype in BOUND_QBRAID_META_TYPES