+++ CHANGELOG.md
@@ -41,6 +41,52 @@ device.profile.program_spec
 #  <ProgramSpec('builtins.str', 'qasm3')>]
 ```
 
+- Added support for Pasqal devices through the `AzureQuantumProvider` with `pulser` program type ([#947](https://github.com/qBraid/qBraid/pull/947)). For example:
+
+```python
+import numpy as np
+import pulser
+from azure.quantum import Workspace
+from qbraid.runtime import AzureQuantumProvider
+
+connection_string = "[Your connection string here]"
+workspace = Workspace.from_connection_string(connection_string)
+
+provider = AzureQuantumProvider(workspace=workspace)
+
+input_data = {}
+
+qubits = {
+    "q0": (0, 0),
+    "q1": (0, 10),
+    "q2": (8, 2),
+    "q3": (1, 15),
+    "q4": (-10, -3),
+    "q5": (-8, 5),
+}
+register = pulser.Register(qubits)
+
+sequence = pulser.Sequence(register, pulser.DigitalAnalogDevice)
+sequence.declare_channel("ch0", "rydberg_global")
+
+amp_wf = pulser.BlackmanWaveform(1000, np.pi)
+det_wf = pulser.RampWaveform(1000, -5, 5)
+pulse = pulser.Pulse(amp_wf, det_wf, 0)
+sequence.add(pulse, "ch0")
+
+device = provider.get_device("pasqal.sim.emu-tn")
+
+job = device.run(sequence, shots=1)
+
+job.wait_for_final_state()
+
+job.status()  # <COMPLETED: 'job has successfully run'>
+
+result = job.result()
+
+result.data.get_counts()  # {'100110': 1}
+```
+
 ### Improved / Modified
 - Prepped tests for supporting `qiskit>=2.0` ([#955](https://github.com/qBraid/qBraid/pull/955))